+++ CHANGELOG.md
@@ -8,10 +8,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 #### Added
 
+- Added basic regression tests [#425](https://github.com/askap-vast/vast-pipeline/pull/425)
+- Added image length validation for config [#425](https://github.com/askap-vast/vast-pipeline/pull/425)
+
 #### Changed
 
 #### Fixed
 
+- Fixed config files in tests [#430](https://github.com/askap-vast/vast-pipeline/pull/430)
 - Fixed sources table on measurement detail page [#429](https://github.com/askap-vast/vast-pipeline/pull/429).
 - Fixed missing meta columns in parallel association [#427](https://github.com/askap-vast/vast-pipeline/pull/427).
 
@@ -19,6 +23,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 #### List of PRs
 
+- [#430](https://github.com/askap-vast/vast-pipeline/pull/430) fix: Test config files.
+- [#425](https://github.com/askap-vast/vast-pipeline/pull/425) feat: Basic regression tests.
 - [#429](https://github.com/askap-vast/vast-pipeline/pull/429) fix: Fixed sources table on measurement detail page.
 - [#427](https://github.com/askap-vast/vast-pipeline/pull/427) fix: Fixed missing meta columns in parallel association.